+++ lib/src/adaptive_widget.dart
@@ -10,6 +10,18 @@ abstract class AdaptiveWidget extends Widget {
   @override
   AdaptiveElement createElement() => AdaptiveElement(this);
 
+  /// {@macro flutter.widgets.Widget.build}
+  ///
+  /// Called when the widget is inserted into the tree.
+  ///
+  /// Override this method to handle the build process for this widget.
+  ///
+  /// The [framework] argument is the [StatelessElement] object for this
+  /// widget. The [child] argument is the child widget, if any.
+  ///
+  /// Returns the widget itself.
+  ///
+  /// Default widget build of [AdaptiveStateLessWidget]
   @protected
   Widget build(BuildContext context);
 
@@ -17,6 +29,7 @@ abstract class AdaptiveWidget extends Widget {
   ///
   /// Returns the widget that matches the current [BoxConstraints]
   /// (Extra Small, Small, Medium, Large, Extra Large, or XXL).
+  @widgetFactory
   Widget _match(BuildContext context, BoxConstraints constraints) {
     if (constraints.isXS) {
       return xs(context, constraints);
@@ -36,21 +49,38 @@ abstract class AdaptiveWidget extends Widget {
   @nonVirtual
   Widget get adaptive => LayoutBuilder(builder: _match);
 
+  /// The xs method is called when the screen size is XS or larger.
+  /// By default, it calls the build method.
   @protected
   Widget xs(BuildContext context, BoxConstraints constraints) => build(context);
 
+  /// The sm method is called when the screen size is SM or larger.
+  /// By default, it calls the xs method.
   @protected
-  Widget sm(BuildContext context, BoxConstraints constraints) => xs(context, constraints);
+  Widget sm(BuildContext context, BoxConstraints constraints) =>
+      xs(context, constraints);
 
+  /// The md method is called when the screen size is MD or larger.
+  /// By default, it calls the sm method.
   @protected
-  Widget md(BuildContext context, BoxConstraints constraints) => sm(context, constraints);
+  Widget md(BuildContext context, BoxConstraints constraints) =>
+      sm(context, constraints);
 
+  /// The lg method is called when the screen size is LG or larger.
+  /// By default, it calls the md method.
   @protected
-  Widget lg(BuildContext context, BoxConstraints constraints) => md(context, constraints);
+  Widget lg(BuildContext context, BoxConstraints constraints) =>
+      md(context, constraints);
 
+  /// The xl method is called when the screen size is XL or larger.
+  /// By default, it calls the lg method.
   @protected
-  Widget xl(BuildContext context, BoxConstraints constraints) => lg(context, constraints);
+  Widget xl(BuildContext context, BoxConstraints constraints) =>
+      lg(context, constraints);
 
+  /// The xxl method is called when the screen size is XXL or larger.
+  /// By default, it calls the xl method.
   @protected
-  Widget xxl(BuildContext context, BoxConstraints constraints) => xl(context, constraints);
+  Widget xxl(BuildContext context, BoxConstraints constraints) =>
+      xl(context, constraints);
 }