+++ CHANGELOG.md
@@ -1,3 +1,7 @@
 ## 0.0.1
 
-* TODO: Describe initial release.
+* AdaptiveWrapState: Adaptive State for Stateful Widget.
+* AdaptiveStateLessWidget: Adaptive Stateless Widget.
+* AdaptiveWidget: Adaptive Widget with BoxConstrains values for Stateless Widget.
+* AdaptiveWrapWidget: Adaptive Widget with BoxConstrains values for Wrap build method on Stateless Widget.
+* Adaptive Default Setter.